+++ examples/scripts/DFN_size_distributions.py
@@ -49,7 +49,7 @@ output_variables = [
     "Current [A]",
     "X-averaged negative area-weighted particle-size distribution [m-1]",
     "X-averaged positive area-weighted particle-size distribution [m-1]",
-    "Terminal voltage [V]",
+    "Voltage [V]",
 ]
 sims[0].plot(output_variables)
 
@@ -58,6 +58,6 @@ output_variables = [
     "Negative particle surface concentration [mol.m-3]",
     "Positive particle surface concentration [mol.m-3]",
     "Current [A]",
-    "Terminal voltage [V]",
+    "Voltage [V]",
 ]
 pybamm.dynamic_plot(sims, output_variables=output_variables)