+++ pybamm/models/submodels/current_collector/effective_resistance_current_collector.py
@@ -261,7 +261,7 @@ class EffectiveResistance(BaseEffectiveResistance):
             "Negative current collector potential [V]": phi_s_cn,
             "Positive current collector potential [V]": phi_s_cp,
             "Local current collector potential difference [V]": V_cc,
-            "Terminal voltage [V]": V,
+            "Voltage [V]": V,
         }
         return processed_vars
 
@@ -388,6 +388,6 @@ class AlternativeEffectiveResistance2D(BaseEffectiveResistance):
             "Negative current collector potential [V]": phi_s_cn,
             "Positive current collector potential [V]": phi_s_cp,
             "Local current collector potential difference [V]": V_cc,
-            "Terminal voltage [V]": V,
+            "Voltage [V]": V,
         }
         return processed_vars