+++ pybamm/models/submodels/electrolyte_conductivity/base_electrolyte_conductivity.py
@@ -188,6 +188,13 @@ class BaseElectrolyteConductivity(pybamm.BaseSubModel):
             f"{Domain} electrode surface potential difference [V]": delta_phi,
         }
 
+        if Domain == "Negative":
+            variables["Anode potential [V]"] = pybamm.boundary_value(delta_phi, "right")
+        elif Domain == "Positive":
+            variables["Cathode potential [V]"] = pybamm.boundary_value(
+                delta_phi, "left"
+            )
+
         return variables
 
     def _get_electrolyte_overpotentials(self, variables):