+++ pybamm/models/submodels/interface/kinetics/diffusion_limited.py
@@ -43,7 +43,7 @@ class DiffusionLimited(BaseInterface):
         # Get exchange-current density
         j0 = self._get_exchange_current_density(variables)
         # Get open-circuit potential variables and reaction overpotential
-        ocp = variables[f"{Domain} electrode {reaction_name}open circuit potential [V]"]
+        ocp = variables[f"{Domain} electrode {reaction_name}open-circuit potential [V]"]
         eta_r = delta_phi_s - ocp
 
         # Get interfacial current densities