+++ pybamm/models/submodels/interface/open_circuit_potential/single_ocp.py
@@ -37,23 +37,34 @@ class SingleOpenCircuitPotential(BaseOpenCircuitPotential):
                     sto_surf = sto_surf.orphans[0]
                     T = T.orphans[0]
 
-            ocp = self.phase_param.U(sto_surf, T)
+            ocp_surf = self.phase_param.U(sto_surf, T)
             dUdT = self.phase_param.dUdT(sto_surf)
+
+            # Bulk OCP is from the average SOC and temperature
+            sto_bulk = variables[f"{Domain} electrode{phase_name} stoichiometry"]
+            T_bulk = pybamm.xyz_average(T)
+            ocp_bulk = self.phase_param.U(sto_bulk, T_bulk)
         elif self.reaction == "lithium metal plating":
             T = variables[f"{Domain} electrode temperature [K]"]
-            ocp = 0 * T
+            ocp_surf = 0 * T
+            ocp_bulk = 0 * T
             dUdT = 0 * T
         elif self.reaction == "lead-acid main":
             c_e = variables[f"{Domain} electrolyte concentration [mol.m-3]"]
             # If c_e was broadcast, take only the orphan
             if isinstance(c_e, pybamm.Broadcast):
                 c_e = c_e.orphans[0]
-            ocp = self.phase_param.U(c_e, self.param.T_init)
+            ocp_surf = self.phase_param.U(c_e, self.param.T_init)
             dUdT = pybamm.Scalar(0)
+
+            # Bulk OCP is from the average concentration and temperature
+            c_e_av = variables["X-averaged electrolyte concentration [mol.m-3]"]
+            ocp_bulk = self.phase_param.U(c_e_av, self.param.T_init)
 
         elif self.reaction == "lead-acid oxygen":
-            ocp = self.param.U_Ox
+            ocp_surf = self.param.U_Ox
+            ocp_bulk = self.param.U_Ox
             dUdT = pybamm.Scalar(0)
 
-        variables.update(self._get_standard_ocp_variables(ocp, dUdT))
+        variables.update(self._get_standard_ocp_variables(ocp_surf, ocp_bulk, dUdT))
         return variables