+++ pybamm/models/submodels/particle/fickian_diffusion.py
@@ -239,7 +239,6 @@ class FickianDiffusion(BaseParticle):
             variables.update(self._get_standard_diffusivity_variables(D_eff))
 
         variables.update(self._get_standard_flux_variables(N_s))
-        variables.update(self._get_total_concentration_variables(variables))
 
         return variables