+++ pybamm/models/submodels/particle/x_averaged_polynomial_profile.py
@@ -219,7 +219,6 @@ class XAveragedPolynomialProfile(PolynomialProfile):
             )
         )
         variables.update(self._get_standard_flux_variables(N_s))
-        variables.update(self._get_total_concentration_variables(variables))
 
         return variables