+++ tests/unit/test_expression_tree/test_operations/test_latexify.py
@@ -28,7 +28,7 @@ class TestLatexify(unittest.TestCase):
         self.assertIn(r"\\", str(model_spme.latexify(newline=False)))
 
         # Test terminal voltage equation name
-        self.assertIn("Terminal voltage [V]", func_spme)
+        self.assertIn("Voltage [V]", func_spme)
 
         # Test partial derivative in boundary conditions
         self.assertIn("partial r}", func_spme)